cinny-desktop: fix build failure
Unfortunately this leads to the loss of the system tray icon and the
functionality that comes with that. I was unfortunately unable to figure
out why this was the case, but in response, I have removed the
dependencies that were included for it.

Hopefully somebody more experienced than me can fix it.

Fixes #357406

(cherry picked from commit 7d6d354a153530d950acbe126809653a04963534)