darwin.stdenv: use native Clang as a cross-compiler
In theory, Clang is a cross-compiler by default. After the following
PRs, this is true on platforms other than Darwin.

- https://github.com/NixOS/nixpkgs/pull/355532
- https://github.com/NixOS/nixpkgs/pull/356162

The reason Darwin needs additional changes is it only overlays Clang and
LLVM in the native target case. When cross-compiling, it builds Clang
and LLVM again. This is unnecessary as long as the Darwin stdenv avoids
propagating the Clang wrappers, which are target-dependent.

In theory, ld64 is also a cross-linker by default. Realizing that for
nixpkgs requires additional work that will be done separately.