ocamlPackages.menhir: support --suggest-menhirLib
menhir provides a `--suggest-menhirLib` option that tries to infer the
path of the menhir library from the path of the menhir binary.

Since the menhir library and the menhir binary are built as different
derivations, this does not work.

This patch hardcodes the location of the menhir library into the menhir
binary, making `--suggest-menhirLib` work.