Revert "nixos/frigate: provide ffmpeg-full for nvidia hw accel"
With https://github.com/NixOS/nixpkgs/pull/354952, the regular ffmpeg-headless
should now have all the deps required for frigate to make use of hardware
acceleration on all hardware.

This is progress towards https://github.com/NixOS/nixpkgs/issues/271863

This reverts commit 7e33e470df5dcbb9c84d6a08a1128953c4281e17.

This may be surprising as it's quite soon after
7e33e470df5dcbb9c84d6a08a1128953c4281e17 was merged but we decided in
https://github.com/NixOS/nixpkgs/pull/357717#discussion_r1853840512 to merge it
quickly to unbreak master and backport to 24.11. Non-full ffmpeg can be pursued
in master using a simple revert which is where we are now.