adminer-pematon: init at 4.12
This is an active fork of adminer by pematon. Seems to be in rapid development.
The index.php is required for plugins to work, and the upstream provides an example, but not a ".php" file.
We need to modify it for NixOS use anyway.

Co-authored-by: Pol Dellaiera <pol.dellaiera@protonmail.com>
Signed-off-by: John Titor <50095635+JohnRTitor@users.noreply.github.com>
(cherry picked from commit 33f90436f53627299d550fc1df4a71742d6d5ca1)