chromium: use newer rustc 1.77 -> 1.80 in preparation of M131
chromium expects nightly/bleeding edge rustc features which we enable in
our stable rustc release by setting RUSTC_BOOTSTRAP=1.

The default rustc in 24.05 however is too old even with that workaround,
but thankfully we also have 1.80 as opt-in. So we use that.

This is slightly closer to the rustc version we have in unstable (1.82)
and unbreaks building the upcoming M131.

Previously: bad7d0f1fabaa4dd0dc6a97a51fd2ac1efe5d9dd