aegisub: switch to arch1t3cht’s fork
Aegisub is a program of many forks. We currently use the inactive
wangqr fork that has not received an update in two years. This
switches to the actively‐maintained arch1t3cht fork, bringing in
additional functionality, bug fixes, and a much more reasonable Meson
build system that lets us drop our LuaJIT devendoring patch. It also
contains fixes for newer versions of Boost.

Note that although the `README.md` has a note from two years ago saying
“**Don't** use this version if you're just looking for any version
of Aegisub”, that is somewhat outdated and predates it becoming a
fully‐fledged fork: as the next section of the file from two months
ago admits, it is the current de facto standard Aegisub fork and the
version universally recommended by up‐to‐date sources; see e.g.:

* <https://fansubbers.miraheze.org/wiki/Aegisub>
* <https://guide.encode.moe/typesetting/aegisub.html>
* <https://www.reddit.com/r/aegisub/comments/yzvi4m/aegisub_update_status/>
* <https://github.com/Ristellise/AegisubDC/issues/10>