osrm-backend: 5.27.1 -> 5.27.1-unstable-2024-11-03
Includes fixes for newer versions of GCC and Boost and obsoletes our
patch and most of our flags.

Upstream has not cut a release tag in over two years. They have
signalled a desire to make a new release this year, but have limited
resources to do so:

* <https://github.com/Project-OSRM/osrm-backend/issues/6747>
* <https://github.com/Project-OSRM/osrm-backend/issues/6778>
* <https://github.com/Project-OSRM/osrm-backend/issues/7028>

The NixOS test passes, so I think this should be okay and leave this
package in a healthier state than it is currently in.