litecoin{,d}: drop
Our current version is from March 2023; there was a release with a
critical security fix in March of this year, and another one with more
security fixes two weeks ago. The listed maintainer has been inactive
in Nixpkgs for years and it seems nobody else is stepping up to keep
this package secure and updated. Let’s drop it until that changes.