libbitcoin{,-client,-explorer,-network,-protocol}: drop
These are the last packages in the tree that depend on Boost
1.75. There hasn’t been a release since 2023; upstream have stated
in <https://github.com/libbitcoin/libbitcoin-system/issues/1234> that
they do not plan to maintain the current version, and are currently
working on a backwards‐incompatible major update.

Nothing in the tree uses these libraries. The only other software
distributions on Repology packaging them are YACP and the AUR, neither
of which has had an update since 2019. They were removed from Ubuntu
after 16.04, and have more recently been removed from [SlackBuilds.org]
and [Homebrew] with the same reasoning that they’re keeping Boost
1.75 around unnecessarily and no packaged software wants them.

[SlackBuilds.org]: https://github.com/SlackBuildsOrg/slackbuilds/commit/17c7d0e430eadaab2ea02d435b0f48977961eba9
[Homebrew]: https://github.com/Homebrew/homebrew-core/pull/157427

Since these packages are also unmaintained in Nixpkgs, let’s just
drop them and if anyone wants to package the 4.x version when it’s
out they can re‐add it.