libstdcxx5: remove
I observed this package to be a big red herring when people are searching
for the package they need if libstdc++ is missing: I observed this to
happen if you're either very new to packaging and don't know where it's
supposed to come from or if you package some binary program (or virtual
environment) and this library is missing.

The package is a subset of GCC 3.3 from 2005 and only needed for very
old proprietary software that cannot be recompiled where it's only
questionable if this is even appropriate to have in nixpkgs.

There's in fact one such package, unreal tournament 2004. As much as I
don't think that libstdcxx5 has a place in here, this also applies to
it. If people want to consume this, it should belong into an external
flake.

Until this happens, I'm sorry for https://www.youtube.com/watch?v=mXapt4GHt-s