vimPlugins: Remove the nvim-cmp dependency on cmp-sources
The `blink.cmp` neovim plugin provides a shim to use sources initially
developped for `nvim-cmp` by using the `blink.compat` plugin.

This plugin implements a `cmp` lua module, so that `nvim-cmp` sources
register with `blink.cmp` instead of registering with `nvim-cmp`.

If we have a hard dependency on nvim-cmp for cmp sources this means we
can't use them with `blink.cmp`

This also means that we can't easily require the plugin in the check
phase, as we would need to add nvim-cmp (or blink.compat) to provide the
'cmp' lua module.