buffybox: init at 3.2.0-unstable-2024-11-10
[buffybox](https://gitlab.postmarketos.org/postmarketOS/buffybox/) is a
suite of graphical applications for the terminal.
notably for nixpkgs, this includes `buffyboard`: an on-screen
framebuffer keyboard especially useful for touch-only devices like
mobile phones.

`buffyboard` can be built and launched (as root) with no arguments: then
tab to an unused TTY (e.g. Ctrl+Alt+F2) where an on-screen keyboard should
now be visible (if not, try clicking in the bottom of the terminal to force a redraw).

further configuration of buffyboard is possible by editing
`/etc/buffyboard.conf` or passing CLI flags.