octavePackages: Ensure locales are found and present in final env
Add locale as buildInput to Octave packages buildEnv

Some Octave packages use a Perl script to translate their
documentation to other languages, which shells out to the locale
program. Before, this would error every time locale was called (which
is on every file), and would lead to a large number of what looked
like errors being printed out.

Set the path to the glibc locales using LOCALE_ARCHIVE in the
makeWrapper invocation, so the wrapped program knows where to find
locales for translations at run-time.