darwin.ICU: enable the C++ API by default
The ICU C++ API is not ABI stable. Since ICU is normally shipped as a
system library, Apple has disabled it by default. This breaks using ICU
as the default implementation on Darwin in nixpkgs because packages
expect to use the C++ API without having to opt into it.

This change enables the C++ API by default based on the assumption that
the packages using it in nixpkgs are built from source, and any binaries
that would link against `libicucore.B.dylib` won’t be using the C++ API
and should not suffer from ABI problems, and if they did use the C++
API, they would have ABI problems regardless because the symbols are
still present in the ICU dylibs.