darwin.ICU: use char16_t instead of uint16_t in the C++ API
This matches the default configuration of ICU upstream. Otherwise, some
packages (such as Boost) fail to build against darwin.ICU. This should
be safe ABI-wise because (according to the comments), char16_t and
uint16_t have the same memory representation. The reason why Apple did
not make the change is due to a build failure in one of their
proprietary frameworks when using char16_t.