aflplusplus: add a knob for persistent replay support
This adds a trigger for enabling replaying of persistent mode crashes with AFL++. This can be useful if the target is stateful and crashes aren't reproducible on their own. Due to the state-keeping, this comes at a performance cost and thus is disabled by default. To use it, AFL++ needs to be started with the `AFL_PERSISTENT_RECORD` environment variable set to a number [1].

[1]: https://github.com/AFLplusplus/AFLplusplus/blob/3af042d5bf3116a194168f5d1e40311a484959e8/include/config.h#L92-L98