llvmPackages: Drop buildLlvmTools from llvm package derivations
This an RFC in the form of a patch.

I've noticed that if you try to build a (buildPlatform != hostPlatform)
llvmPackages which aren't available via hydra, you incur a build of
LLVM.

(For example, pkgsStatic.llvmPackages_git.llvm does two builds of LLVM).

LLVM has the capability of cross-building an llvm-tblgen internally. And
tablegen is relatively cheap to rebuild, only requiring 8 CPU minutes or
so to build. LLVM is much more expensive to build.

So I propose to drop LLVM_TABLEGEN_EXE and related variables and let the
LLVM build system do its thing. It does mean that a buildPlatform
llvm-tblgen will get built (and discarded) repeatedly in different
derivations (e.g. in .llvm and .clang), but at least it is less
expensive than building LLVM and Clang twice.

Note that I haven't interfered with the cross compiler wrapper logic,
I think buildLlvmTools must remain there, only the llvm packages
themselves.

Signed-off-by: Peter Waller <p@pwaller.net>