ldmud: use libxcrypt-legacy
The LDMud game driver defaults to using a legacy DES hash that is
(sensibly) only enabled in the libxcrypt-legacy derivation, not
libxcrypt.

This commit switches LDMud to that build input instead of libxcrypt.
This fixes a runtime error when calling `efun::crypt()` from LPC that
would print a confusing error message like "crypt() is not available.".