gnomeExtensions.argos: Update to more recent revision with support for GNOME 47
The current extensions does not work with GNOME 47, which the new release nixos-24.11 is based. The above is an update to the latest commit which enables support for GNOME 47.

See:
https://github.com/p-e-w/argos/commit/cd0de7c79072979bed41e0ad75741bbd8e113950

I tested on my machine and the extension works.

(cherry picked from commit b803e1b03dc698150dff6e856bf7bb22e8751955)