linux-kernels: add LTS flag
This serves to differentiate between a kernel that is an LTS release and one
that is merely a regular "stable" release.

This is required for the ZFS kernel module to show a warning when not used in
combination with an LTS kernel.

As this is a non-breaking change, this will be backported to all supported
channels to allow for simpler backports of updates.