nixos/prometheus: add fallback_scrape_protocol and scrape_protocols options
I chose to not constrain them to the supported values to be more flexible in supporting any
newly added values. But maybe it makes sense to constrain them, I'm not completely sure.

Ran into this because Prometheus 3.0 was introduced last week and my config broke,
since rtl_433 doesn't set the headers / content-type correctly.

Related-to: https://github.com/NixOS/nixpkgs/pull/358862
Related-to: https://github.com/prometheus/prometheus/pull/15136