busybox-sandbox-shell: fix eval for musl
Setting useMusl all the time to a condition dependent on isGnu meant
that for musl platforms, useMusl ended up being set to false, tripping
the assertion in the busybox expression.  The correct thing to do is
to leave useMusl at its default value, except for in the specific case
where we want to force it to true.

Fixes: 11472f0d1b10 ("busybox-sandbox-shell: replace pkgsStatic with useMusl")