authentik: flag with `knownVulnerabilities`
The bump to 2024.10.x is currently stalled in #345940.

The fix for CVE-2024-52289 involves a backward incompatible
API and DB change so it is also not great for a backport even
if we cherry-picks the security fixes.

Given no NixOS module is available in nixpkgs marking the package
with `knownVulnerabilities` should have a limited impact.

(cherry picked from commit 384f6f592b99f7ed384da7282d6aa6166eae075b)