maintainers: update GitHub names
It is important to have a usable mapping of maintainers -> GH handles
to ping maintainers for reviews or bugs.

This PR updates all outdated GitHub names based on the provided ID
(note: I only fixed the casing up to JanLikar).
The script is added in a follow-up commit.

For the affected entries, I checked that the ID is correct by looking
at one of their PRs.