nixos/zfs: skip scrub if one is in progress
The `zpool-scrub` command has only one non-zero return code and that is
for everything: failed to initiate a scrub for _some_ reason, but not
for which reason. Since that doesn't exist, we can check the output of
the `zpool-status` command and see if a scrub is in progress. If it is,
we skip scrubbing the zpool because doing so would do nothing but cause
the `zpool-scrub` command to erroneously exit.

A `for` loop was added to iterate over multiple zpools, in case there
are more than one zpool that the user hasn't specified in the NixOS
configuration.

Fixes #361952.