buildBazelPackage: use same bazel for deps & build
It should make more sense to use same `bazel` to produce & consume the deps.

Produced deps tarball seems to be currently the same if build
with or without `enableNixHacks`, at least have verified it for `fcitx5-mozc`.

Since we also start using `bazelForBuild.name` for version string it
will invalidate the deps, which sounds like a good thing to do on
`buildBazelPackage` changes. This will cause a rebuild but relatively
few packages are affected.