chromium: remove ofborg maintainer ping workaround, use CODEOWNERS
The workaround to have ofborg ping chromium and ungoogled-chromium
maintainers when a change was only made to the upstream-info relied on
string context.

That string context was provided by the upstream-info being a nix file,
not a json file, and then holding on to that string context using
awkward attribute merges.

It was intended as a quick fix until the handling of this would improve
in ofborg itself and worked great.

That was until very recently when we switched from the chromium release
tarball to git source fetching in 8dd2f1add978a4747a5962f2874b8ad20f86b01c.

Part of that change included going back from upstream-info.nix to
upstream-info.json and with that losing the string context and the base
on which this workaround used to work.

But this is fine. A lot has happened in the meantime.

CODEOWNERS was reimplemented and no longer requires every user listed in
it to have write permissions to the repository (commit bit).

Meaning we can accept that ofborg pings no longer work and instead rely
on CODEOWNERS exclusively.

It should, however, be noted that CODEOWNERS provide less granularity
than ofborg, meaning we can no longer differentiate between
ungoogled-chromium and chromium or even chromedriver.

Previously, implementing the workaround that is now essentially
reverted: 68c59791fb6644ac733d99d0147b09bce4cb8319

(cherry picked from commit 674835a9d98fd037c8f6ce1c954b5e07e56b0832)