golden-cheetah-bin: fix qt startup
Due to a version missmatch between the packaged
Qt package from upstream and NixOS Qt version
GoldenCheetah will not be able to start with the following
error message on KDE:

```
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: eglfs, linuxfb, minimal, minimalegl, offscreen, vnc, xcb.
```

This is analog to other errors and is caused by a leaked QT_PLUGIN_PATH
variable from the host to the AppImage[1].

Since this only causes issues on KDE it is sufficient to empty the
variable from the desktop Menu item. This does not fix startup from
the command line and is rather a workaround than a true bugfix and the
underlying issue (leaking variables to appimageTools.wrapType2
applications) should be fixed.

This does not affect other desktop environments

[1]: https://discourse.nixos.org/t/cannot-run-ida-free-with-steam-run-could-not-load-the-qt-platform-plugin-xcb-in/29783

Signed-off-by: Florian Brandes <florian.brandes@posteo.de>