github-runner: use finalAttrs to make it possible to override the version (#363733)
With rec, it is nearly impossible to override the version since it is
used in so many different places in the nix expression.
With the current change, it is possible to override version and src to
build a different version.

(cherry picked from commit d298ddec02bad2108412ffc0f7156c1d91993557)