edk2-uefi-shell: fix cross compilation
This PYTHON_COMMAND workaround was introduced in #201095, but it does
not seem to have caused issues so far, until the version update in
bef3f40f05fbea6154967ff638d4e194f1a7d251. Not sure why this was not an
issue before this update and why it is now, but I think this fix makes
sense regardless of the exact root cause, and I don't think it's worth
the effort to dig further.