workflows: Condition all merge-dependent workflows on a merge commit
After the introduction of the reusable get-merge-commit.yml workflow,
this now applies it on all remaining workflows that depend on a merge
commit. This ensures that CI doesn't fail for those PRs, which would
cause notifications to be sent, and the CI run to be marked as failed in
the list of runs.

(cherry picked from commit 83d4c9d28db493e2b11d6684858966acf137752e)